AMLS-4675 | Further bug fix for NoSubject notifications
