+++ app/views/notifications/your_messages.scala.html
@@ -16,6 +16,7 @@
 
 @import include._
 @import models.notifications._
+@import utils.ContactTypeHelper
 
 @(businessName: String, currentApplicationNotifications: Seq[NotificationRow], previousApplicationNotifications: Seq[NotificationRow] = Seq())(implicit request: Request[_],m:Messages)
 
@@ -42,7 +43,7 @@
             @notifications.map { comms =>
             <tr class="@if(!comms.isRead){unread}">
                 <td>@if(!comms.isRead){<span class="circle"></span><span class="visuallyhidden">Unread</span>}else{<span class="visuallyhidden">Read</span>}</td>
-                <td><a href="@controllers.routes.NotificationController.messageDetails(comms._id.id, comms.getContactType, comms.amlsRegistrationNumber, comms.templatePackageVersion)">@Messages(comms.subject)</a></td>
+                <td><a href="@controllers.routes.NotificationController.messageDetails(comms._id.id, ContactTypeHelper.getContactType(comms.status, comms.contactType, comms.variation), comms.amlsRegistrationNumber, comms.templatePackageVersion)">@Messages(comms.subject)</a></td>
                 <td>@Messages(comms.notificationType)</td>
                 <td>@comms.dateReceived</td>
             </tr>
