+++ test/models/notifications/NotificationRowSpec.scala
@@ -16,7 +16,6 @@
 
 package models.notifications
 
-import models.notifications.ContactType.NoSubject
 import org.joda.time.{DateTime, DateTimeZone}
 import org.scalatestplus.play.PlaySpec
 import play.api.libs.json._
@@ -101,88 +100,6 @@ class NotificationRowSpec extends PlaySpec with AmlsSpec {
         """.stripMargin)
 
       NotificationRow.format.reads(json) must be(JsSuccess(model))
-    }
-
-    "return correct ContactType" when {
-
-      "contact type is populated" in {
-        val model = NotificationRow(
-          Some(
-            Status(
-              Some(StatusType.Revoked),
-              Some(RevokedReason.RevokedCeasedTrading)
-            )),
-          Some(ContactType.MindedToRevoke),
-          None,
-          false,
-          new DateTime(1479730062573L, DateTimeZone.UTC),
-          false,
-          "XJML00000200000",
-          "1",
-          new IDType("5832e38e01000001005ca3ff"
-          ))
-
-        model.getContactType mustBe (ContactType.MindedToRevoke)
-      }
-
-      "auto rejected" in {
-        val model = NotificationRow(
-          Some(
-            Status(
-              Some(StatusType.Rejected),
-              Some(RejectedReason.FailedToPayCharges)
-            )),
-          None,
-          None,
-          false,
-          new DateTime(1479730062573L, DateTimeZone.UTC),
-          false,
-          "XJML00000200000",
-          "1",
-          new IDType("5832e38e01000001005ca3ff"
-          ))
-
-        model.getContactType mustBe (ContactType.ApplicationAutorejectionForFailureToPay)
-      }
-
-      "variation approved" in {
-        val model = NotificationRow(
-          Some(
-            Status(
-              Some(StatusType.Approved), None
-            )),
-          None,
-          None,
-          true,
-          new DateTime(1479730062573L, DateTimeZone.UTC),
-          false,
-          "XJML00000200000",
-          "1",
-          new IDType("5832e38e01000001005ca3ff"
-          ))
-
-        model.getContactType mustBe (ContactType.RegistrationVariationApproval)
-      }
-
-      "DeRegistrationEffectiveDateChange" in {
-        val model = NotificationRow(
-          Some(
-            Status(
-              Some(StatusType.DeRegistered), None
-            )),
-          None,
-          None,
-          true,
-          new DateTime(1479730062573L, DateTimeZone.UTC),
-          false,
-          "XJML00000200000",
-          "1",
-          new IDType("5832e38e01000001005ca3ff"
-          ))
-
-        model.getContactType mustBe (ContactType.DeRegistrationEffectiveDateChange)
-      }
-
     }
 
     "return application failure subject line" when {
@@ -293,27 +210,6 @@ class NotificationRowSpec extends PlaySpec with AmlsSpec {
             ))
         ).subject must be("notifications.rejr.title")
       }
-
-    }
-
-    "return NoSubject when ContactType not determined" in {
-      val model = NotificationRow(
-        Some(
-          Status(
-            Some(StatusType.Approved), None
-          )),
-        None,
-        None,
-        false,
-        new DateTime(1479730062573L, DateTimeZone.UTC),
-        false,
-        "XJML00000200000",
-        "1",
-        new IDType("5832e38e01000001005ca3ff"
-        ))
-
-      model.getContactType mustBe NoSubject
-      model.subject mustBe "notifications.subject.NoSubject"
 
     }
   }
